removed helpers, added comments
